Set the right socket for each client on the server
